requirements.txt
discord_py_interactions==4.3.2
